Refactor: remove inflateSourceBoard and introduce inflatePcbBoard for cleaner PCB handling
